feat: スマホ向けにコピー用の固有 ID を送れるようにうする
